+++ crates/vertigo-cli/src/serve/mod.rs
@@ -5,7 +5,13 @@ mod response_state;
 mod serve_opts;
 mod serve_run;
 mod server_state;
+mod vertigo_handler;
+mod vertigo_install;
 mod wasm;
 
+pub use mount_path::{MountConfig, MountConfigBuilder};
 pub use serve_opts::{ServeOpts, ServeOptsInner};
 pub use serve_run::run;
+pub use server_state::ServerState;
+pub use vertigo_handler::vertigo_handler;
+pub use vertigo_install::vertigo_install;